# openapi.json
{
  "openapi": "3.1.0",
  "info": { "title": "FastAPI", "version": "0.1.0" },
  "paths": {
...
    "/recipes": {
      "get": {
        "tags": ["meal planning"],
        "summary": "Get Recipes",
        "operationId": "get_recipes_recipes_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          }
        }
      },
      "post": {
        "tags": ["meal planning"],
        "summary": "Add Recipe",
        "operationId": "add_recipe_recipes_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AddRecipe" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/recipes/{recipe_id}": {
      "get": {
        "tags": ["meal planning"],
        "summary": "Get Recipe By Id",
        "operationId": "get_recipe_by_id_recipes__recipe_id__get",
        "parameters": [
          {
            "name": "recipe_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "title": "Recipe Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["meal planning"],
        "summary": "Delete Recipe",
        "operationId": "delete_recipe_recipes__recipe_id__delete",
        "parameters": [
          {
            "name": "recipe_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "title": "Recipe Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/week_plan/single_meal": {
      "get": {
        "tags": ["meal planning"],
        "summary": "Get Specific Meal",
        "operationId": "get_specific_meal_week_plan_single_meal_get",
        "parameters": [
          {
            "name": "day",
            "in": "query",
            "required": true,
            "schema": { "type": "string", "title": "Day" }
          },
          {
            "name": "day_time",
            "in": "query",
            "required": true,
            "schema": { "type": "string", "title": "Day Time" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/planner_settings": {
      "get": {
        "tags": ["meal planning"],
        "summary": "Get Planner Settings",
        "operationId": "get_planner_settings_planner_settings_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          }
        }
      },
      "post": {
        "tags": ["meal planning"],
        "summary": "Put Planner Settings",
        "operationId": "put_planner_settings_planner_settings_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PlannerSettings" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/week_plan": {
      "get": {
        "tags": ["meal planning"],
        "summary": "Get Week Plan",
        "operationId": "get_week_plan_week_plan_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          }
        }
      },
      "post": {
        "tags": ["meal planning"],
        "summary": "Post Week Plan",
        "operationId": "post_week_plan_week_plan_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AddWeekPlan" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["meal planning"],
        "summary": "Delete Meal Slot",
        "operationId": "delete_meal_slot_week_plan_delete",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/MealSlot" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/week_plan/replace_meal": {
      "post": {
        "tags": ["meal planning"],
        "summary": "Replace Meal",
        "operationId": "replace_meal_week_plan_replace_meal_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/MealSlot" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AddRecipe": {
        "properties": {
          "name": { "type": "string", "title": "Name" },
          "emoji": { "type": "string", "title": "Emoji" },
          "baseTime": { "type": "integer", "title": "Basetime" },
          "baseServings": {
            "anyOf": [{ "type": "integer" }, { "type": "null" }],
            "title": "Baseservings",
            "default": 2
          },
          "tags": {
            "anyOf": [
              { "items": { "type": "string" }, "type": "array" },
              { "type": "null" }
            ],
            "title": "Tags",
            "default": []
          },
          "favorited": {
            "anyOf": [{ "type": "boolean" }, { "type": "null" }],
            "title": "Favorited",
            "default": false
          },
          "ingredients": {
            "items": { "$ref": "#/components/schemas/Ingredient" },
            "type": "array",
            "title": "Ingredients"
          },
          "steps": {
            "items": { "type": "string" },
            "type": "array",
            "title": "Steps"
          }
        },
        "type": "object",
        "required": ["name", "emoji", "baseTime", "ingredients", "steps"],
        "title": "AddRecipe"
      },
...
      "AddWeekPlan": {
        "properties": {
          "week": { "$ref": "#/components/schemas/WeekPlan" },
          "DaySettings": { "$ref": "#/components/schemas/WeekSettings" }
        },
        "type": "object",
        "required": ["week", "DaySettings"],
        "title": "AddWeekPlan"
      },
...
      "DayPlan": {
        "properties": {
          "breakfast": {
            "anyOf": [
              { "$ref": "#/components/schemas/MealSlot" },
              { "type": "null" }
            ]
          },
          "lunch": {
            "anyOf": [
              { "$ref": "#/components/schemas/MealSlot" },
              { "type": "null" }
            ]
          },
          "dinner": {
            "anyOf": [
              { "$ref": "#/components/schemas/MealSlot" },
              { "type": "null" }
            ]
          }
        },
        "type": "object",
        "title": "DayPlan"
      },
      "DaySettings": {
        "properties": {
          "day": { "type": "string", "title": "Day" },
          "day_meal_time_type": {
            "$ref": "#/components/schemas/DayType",
            "default": "normal"
          },
          "breakfast_blocked": {
            "type": "boolean",
            "title": "Breakfast Blocked",
            "default": false
          },
          "lunch_blocked": {
            "type": "boolean",
            "title": "Lunch Blocked",
            "default": false
          },
          "dinner_blocked": {
            "type": "boolean",
            "title": "Dinner Blocked",
            "default": false
          }
        },
        "type": "object",
        "required": ["day"],
        "title": "DaySettings"
      },
      "DayType": {
        "type": "string",
        "enum": ["quick", "normal", "relaxed"],
        "title": "DayType"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": { "$ref": "#/components/schemas/ValidationError" },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Ingredient": {
        "properties": {
          "amount": {
            "anyOf": [{ "type": "integer" }, { "type": "null" }],
            "title": "Amount"
          },
          "unit": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Unit"
          },
          "name": { "type": "string", "title": "Name" },
          "count": {
            "anyOf": [{ "type": "integer" }, { "type": "null" }],
            "title": "Count"
          }
        },
        "type": "object",
        "required": ["name"],
        "title": "Ingredient"
      },
...
      "MealSlot": {
        "properties": {
          "recipe_id": { "type": "integer", "title": "Recipe Id" },
          "servings": { "type": "integer", "title": "Servings" },
          "meal_type": { "type": "string", "title": "Meal Type" },
          "day": { "type": "string", "title": "Day" },
          "day_time": { "type": "string", "title": "Day Time" }
        },
        "type": "object",
        "required": ["recipe_id", "servings", "meal_type", "day", "day_time"],
        "title": "MealSlot"
      },
      "PlannerSettings": {
        "properties": {
          "defaultServings": { "type": "integer", "title": "Defaultservings" },
          "quickMealMinutes": {
            "type": "integer",
            "title": "Quickmealminutes"
          },
          "normalMealMinutes": {
            "type": "integer",
            "title": "Normalmealminutes"
          }
        },
        "type": "object",
        "required": [
          "defaultServings",
          "quickMealMinutes",
          "normalMealMinutes"
        ],
        "title": "PlannerSettings"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] },
            "type": "array",
            "title": "Location"
          },
          "msg": { "type": "string", "title": "Message" },
          "type": { "type": "string", "title": "Error Type" },
          "input": { "title": "Input" },
          "ctx": { "type": "object", "title": "Context" }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      },
      "WeekPlan": {
        "properties": {
          "mo": { "$ref": "#/components/schemas/DayPlan", "default": {} },
          "tu": { "$ref": "#/components/schemas/DayPlan", "default": {} },
          "we": { "$ref": "#/components/schemas/DayPlan", "default": {} },
          "th": { "$ref": "#/components/schemas/DayPlan", "default": {} },
          "fr": { "$ref": "#/components/schemas/DayPlan", "default": {} },
          "sa": { "$ref": "#/components/schemas/DayPlan", "default": {} },
          "su": { "$ref": "#/components/schemas/DayPlan", "default": {} }
        },
        "type": "object",
        "title": "WeekPlan"
      },
      "WeekSettings": {
        "properties": {
          "mo": {
            "$ref": "#/components/schemas/DaySettings",
            "default": {
              "day": "monday",
              "day_meal_time_type": "normal",
              "breakfast_blocked": false,
              "lunch_blocked": false,
              "dinner_blocked": false
            }
          },
          "tu": {
            "$ref": "#/components/schemas/DaySettings",
            "default": {
              "day": "tuesday",
              "day_meal_time_type": "normal",
              "breakfast_blocked": false,
              "lunch_blocked": false,
              "dinner_blocked": false
            }
          },
          "we": {
            "$ref": "#/components/schemas/DaySettings",
            "default": {
              "day": "wednesday",
              "day_meal_time_type": "normal",
              "breakfast_blocked": false,
              "lunch_blocked": false,
              "dinner_blocked": false
            }
          },
          "th": {
            "$ref": "#/components/schemas/DaySettings",
            "default": {
              "day": "thursday",
              "day_meal_time_type": "normal",
              "breakfast_blocked": false,
              "lunch_blocked": false,
              "dinner_blocked": false
            }
          },
          "fr": {
            "$ref": "#/components/schemas/DaySettings",
            "default": {
              "day": "friday",
              "day_meal_time_type": "normal",
              "breakfast_blocked": false,
              "lunch_blocked": false,
              "dinner_blocked": false
            }
          },
          "sa": {
            "$ref": "#/components/schemas/DaySettings",
            "default": {
              "day": "saturday",
              "day_meal_time_type": "normal",
              "breakfast_blocked": false,
              "lunch_blocked": false,
              "dinner_blocked": false
            }
          },
          "su": {
            "$ref": "#/components/schemas/DaySettings",
            "default": {
              "day": "sunday",
              "day_meal_time_type": "normal",
              "breakfast_blocked": false,
              "lunch_blocked": false,
              "dinner_blocked": false
            }
          }
        },
        "type": "object",
        "title": "WeekSettings"
      }
    }
  }
}